requirements.txt
attrs==21.4.0
Automat==20.2.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==36.0.1
docutils==0.18.1
hotreload==2020.6.12.15.23.48
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.4
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip
loguru==0.6.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyjnius==1.4.1
pyOpenSSL==22.0.0
python-dotenv==0.19.2
requests==2.27.1
service-identity==21.1.0
six==1.16.0
tomli==2.0.1
Twisted==22.2.0
typing-extensions==4.1.1
urllib3==1.26.8
watchdog==2.1.6
zope.interface==5.4.0
